requirements.txt
allure-pytest==2.13.5
pytest==8.1.1
pytest-playwright==0.5.1
